moved login to / and updated links
