+++ pages/index.js
@@ -1,70 +1,96 @@
 import Head from 'next/head'
 import Image from 'next/image'
-import styles from '../styles/Home.module.css'
+import Link from 'next/link'
+import Router from 'next/router'
+import { withSessionSsr, server } from './lib/config/withSession';
 
+import styles from '../styles/Login.module.css'
+import unfLogo from '../public/UNF_Logo.gif'
+import { useEffect } from 'react';
 
-export default function Home() {
-  return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+export const getServerSideProps = withSessionSsr(
+    async ({req}) => {
+        const user = req.session.user;
 
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
+        if(user) {
+            return {
+                props: { user }
+            }
+        }
 
-        <p className={styles.description}>
-          Get started by editing{' '}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
+        return {
+            props: {}
+        }
+    }
+);
 
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
+const handleSubmit = async (event) => {
+    event.preventDefault();
+  
+    const response = await fetch('/api/sessions', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+            username: event.target.username.value,
+            password: event.target.password.value
+        })
+    });
 
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
+    const result = await response.json();
 
-          <a
-            href="https://github.com/vercel/next.js/tree/canary/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
-  )
+    if (response.ok) {
+        return Router.push("/home");
+    }
+    else {
+        alert(result.message);
+    }
 }
+
+const Login = ({user}) => {
+    useEffect(() => {
+        if (user) {
+            Router.push("/home");
+        }
+    })
+    
+    return (
+        <div>
+            <Head>
+                <title>Login</title>
+                <meta name="description" content="Login page" />
+                <link rel="icon" href="/favicon.ico" />
+            </Head>
+
+            <header className={styles.header}>
+                <Image src={unfLogo} alt="UNF"/>
+            </header>
+
+            <main className={styles.main}>
+
+                <h1 className={styles.description}>
+                    UNF School of Computing <br></br> Honors in Computing Hours Tracking
+                </h1>
+
+                <form className={styles.description} onSubmit={handleSubmit} method="post">
+                    <input name='username' className={styles.input} type="text" placeholder="Username" required/>
+                    <br></br>
+                    <input name='password' className={styles.input} type="password" placeholder="Password" required/>
+                    <br></br>
+                    <input type="submit" value="Log in"/>
+                </form>
+
+                <div className={styles.links}>
+                    <Link href="/username_recovery">Forgot username?</Link>
+                    <span className={styles.linkSpacing}></span>
+                    <Link href="/password_reset">Forgot Password?</Link>
+                    <span className={styles.linkSpacing}></span>
+                    <Link href="/account_creation">Create an account</Link>
+                </div>
+            </main>
+        </div>
+    )
+};
+
+export default Login;
