+++ pages/password_reset.js
@@ -33,7 +33,7 @@ export default function PasswordReset() {
         // Update here
         if (response.status == 200) {
             alert(result.message)
-            Router.push("/login")
+            Router.push("/")
         }
         else if (response.status == 400) {
             alert(result.message)
